+++ packages/sdks/output/svelte/CHANGELOG.md
@@ -1,5 +1,27 @@
 # Builder.io Svelte SDK Changelog (@builder.io/sdk-svelte)
 
+## 3.0.7
+
+### Patch Changes
+
+- 7d01119: feat: Add support for `xsmall` additional breakpoint.
+
+## 3.0.6
+
+### Patch Changes
+
+- abe5cba: Feat: exports `setClientUserAttributes` helper that can be used to set and update Builder's user attributes cookie. This cookie is used by Builder's Personalization Containers to decide which variant to render.
+
+  Usage example:
+
+  ```ts
+  import { setClientUserAttributes } from "@builder.io/sdk-svelte";
+
+  setClientUserAttributes({
+    device: "tablet",
+  });
+  ```
+
 ## 3.0.5
 
 ### Patch Changes